if there is no difference between the old and new data, we skip detach and attach.

+++ src/DataForm/Field/Field.php
@@ -501,6 +501,10 @@ abstract class Field extends Widget
                     $old_data = $this->relation->get()->modelKeys();
                     $new_data = explode($this->serialization_sep, $data);
 
+                    if (count(array_diff($old_data, $new_data)) == 0) {
+                        continue;
+                    }
+
                     $this->relation->detach($old_data);
 
                     if ($data=='') {
